requirements.txt
basicsr==1.4.2
einops==0.8.0
matplotlib==3.7.2
numpy==1.24.4
opencv_contrib_python==4.8.1.78
opencv_python==4.8.0.76
Pillow==10.1.0
Pillow==10.4.0
skimage==0.0
torch==2.1.0+cu121
torchvision==0.16.0+cu121
